优化了图片下载流程，添加LruCache缓存

+++ app/src/main/java/com/luoruiyong/weblog/util/SDUtil.java
@@ -1,8 +1,10 @@
 package com.luoruiyong.weblog.util;
 
+import android.Manifest;
 import android.content.ContentUris;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
@@ -10,7 +12,10 @@ import android.net.Uri;
 import android.os.Build;
 import android.provider.DocumentsContract;
 import android.provider.MediaStore;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
 
+import com.luoruiyong.weblog.base.BaseUi;
 import com.luoruiyong.weblog.base.C;
 import com.luoruiyong.weblog.model.Picture;
 
@@ -19,13 +24,21 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 
 /**内存存储管理，提供给AppCache使用
+ * 本类对外功能：
+ * 1.保存原片到本地
+ * 2.从本地获取原图
+ * 3.从本地获取头像缩略图
+ * 4.从本地获取微博图片缩略图
+ * 5.从本地获取自定义缩略图
+ * 6.获取用户通过调用系统图库所选择图片的路径
+ * 7.清除应用缓存
+ * 8.获取图片缩放比例
+ * 9.获取图片本地存储绝对路径
  * Created by Administrator on 2017/9/7.
  */
 
 public class SDUtil {
     private static final String CLASS_NAME = SDUtil.class.getSimpleName() + "-->";
-    public static final int USER_ICON = 0;
-    public static final int BLOG_IMAGE = 1;
 
     /**
      * 保存图片
@@ -33,22 +46,34 @@ public class SDUtil {
      * @param url  图片资源在服务器中的路径
      * @return  处理结果
      */
-    public static boolean saveImage(Bitmap bitmap , String url, String type){
-        String fileName = AppUtil.md5(url);
-        fileName = getRealFileName(fileName,type);
-        try {
-            File file = new File(fileName);
-            if(file.exists()){
-                file.delete();
+    public static boolean saveImage(Context context,Bitmap bitmap , String url){
+        if(ContextCompat.checkSelfPermission(context, Manifest.permission.WRITE_EXTERNAL_STORAGE)
+                != PackageManager.PERMISSION_GRANTED){
+            ActivityCompat.requestPermissions((BaseUi)context,
+                    new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},BaseUi.REQUEST_PERMISSION_CODE);
+            LogUtil.d(CLASS_NAME+"申请SD卡读写权限");
+        }else{
+            String fileName = getRealFileName(url);
+            String type = getFileType(url);
+            LogUtil.d(CLASS_NAME+"尝试保存图片");
+            try {
+                File file = new File(fileName);
+                if(file.exists()){
+                    file.delete();
+                }
+                FileOutputStream os = new FileOutputStream(file);
+                if (type.equals(Picture.TYPE_PNG)) {
+                    bitmap.compress(Bitmap.CompressFormat.PNG,100,os);
+                }else if(type.equals(Picture.TYPE_JEPG)){
+                    bitmap.compress(Bitmap.CompressFormat.JPEG,100,os);
+                }
+                os.flush();
+                os.close();
+                LogUtil.d(CLASS_NAME+"成功保存原图,路径："+fileName);
+                return true;
+            } catch (IOException e) {
+                LogUtil.d(CLASS_NAME+"保存原图失败" + e.getMessage());
             }
-            FileOutputStream os = new FileOutputStream(file);
-            bitmap.compress(Bitmap.CompressFormat.PNG,100,os);
-            os.flush();
-            os.close();
-            LogUtil.d(CLASS_NAME+"保存图片成功,路径："+fileName);
-            return true;
-        } catch (IOException e) {
-            LogUtil.d(CLASS_NAME+"保存图片失败" + e.getMessage());
         }
         return false;
     }
@@ -56,37 +81,43 @@ public class SDUtil {
     /**
      * 获取图片位图
      * @param url 图片资源在服务器中的路径
-     * @param type   资源类型
      * @return  图片位图或null
      */
-    public static Bitmap getImage(String url,String type){
-        String fileName = AppUtil.md5(url);
-        fileName = getRealFileName(fileName,type);
+    public static Bitmap getOriginalImage(String url){
+        String fileName = getRealFileName(url);
+        LogUtil.d(CLASS_NAME+"尝试从本地获取原图："+fileName);
         File file = new File(fileName);
         if(!file.exists()){
-            LogUtil.d(CLASS_NAME+"从SD卡中获取图片失败");
+            LogUtil.d(CLASS_NAME+"从本地中获取原图失败");
             return null;
         }
-        BitmapFactory.Options options = new BitmapFactory.Options();
-        options.inJustDecodeBounds = false;
-        LogUtil.d(CLASS_NAME+"成功从SD卡中获取图片");
-        return BitmapFactory.decodeFile(fileName,options);
+        LogUtil.d(CLASS_NAME+"成功从本地中获取原图");
+        return BitmapFactory.decodeFile(fileName);
+    }
+
+    //获取头像缩略图
+    public static Bitmap getSampleContactIcon(String url){
+        return getSampleCustomImage(url,Picture.CONTACT_ICON_WIDTH,Picture.CONTACT_ICON_HEIGHT);
+    }
+
+    //获取微博图片缩略图
+    public static Bitmap getSampleBlogImage(String url){
+        return getSampleCustomImage(url,Picture.BLOG_IMAGE_WIDTH,Picture.BLOG_IMAGE_HEIGHT);
     }
 
     /**
-     * 获取图片缩略图，防止出现OOM
+     * 自定义获取图片缩略图
      * @param url  图片资源在服务器中的路径
-     * @param type   资源类型
      * @param width   要求宽度
      * @param height  要求高度
      * @return  缩略图
      */
-    public static Bitmap getSampleImage(String url,String type,int width,int height){
-        String fileName = AppUtil.md5(url);
-        fileName = getRealFileName(fileName,type);
+    public static Bitmap getSampleCustomImage(String url, int width, int height){
+        String fileName = getRealFileName(url);
+        LogUtil.d(CLASS_NAME+"尝试从本地获取缩放图："+fileName);
         File file = new File(fileName);
         if(!file.exists()){
-            LogUtil.d(CLASS_NAME+"从SD卡获取缩放图片失败");
+            LogUtil.d(CLASS_NAME+"从本地获取缩放图失败");
             return null;
         }
         BitmapFactory.Options options = new BitmapFactory.Options();
@@ -95,7 +126,7 @@ public class SDUtil {
         int sampleRatio = getSampleRatio(options.outWidth,options.outHeight,width,height);
         options.inJustDecodeBounds = false;
         options.inSampleSize = sampleRatio;
-        LogUtil.d(CLASS_NAME+"成功从SD卡获取缩放图片");
+        LogUtil.d(CLASS_NAME+"成功从SD卡获取缩放图");
         return BitmapFactory.decodeFile(fileName,options);
     }
 
@@ -153,10 +184,9 @@ public class SDUtil {
      * 清除图片缓存
      * @return  处理结果
      */
-    public static boolean ClearCache(){
-        File imagesDir = new File(C.dir.image);
-        File iconDir = new File(C.dir.icon);
-        if(deleteDirWithFile(imagesDir) && deleteDirWithFile(iconDir)){
+    public static boolean clearCache(){
+        File imagesDir = new File(C.dir.cache);
+        if(deleteDirWithFile(imagesDir)){
             LogUtil.d(CLASS_NAME+"成功清除缓存");
             return true;
         }else{
@@ -188,24 +218,31 @@ public class SDUtil {
 
     /**
      * 获取图片资源的真实路径（目录+文件名），提供给获取图片，保存图片等函数调用
-     * @param fileName  文件名
-     * @param type   资源类型
+     * @param url 图片资源路径
      * @return  真实路径
      */
-    private static String getRealFileName(String fileName,String type){
-        String fileDir = "";
-        if(type.equals(Picture.TYPE_ICON)) {
-            fileDir = C.dir.icon;
-        }else if(type.equals(Picture.TYPE_IMAGE)){
-            fileDir = C.dir.image;
-        }
+    public static String getRealFileName(String url){
+        String fileName = AppUtil.md5(url);
+        String fileDir = C.dir.cache;
         File dir = new File(fileDir);
         if(!dir.exists()){
             dir.mkdirs();
         }
-        fileName = fileDir + fileName + ".png";
-        LogUtil.d(CLASS_NAME+"获得图片文件绝对路径："+fileName);
+        fileName = fileDir + fileName;
+        LogUtil.d(CLASS_NAME+"获取到的图片文件绝对路径："+fileName);
         return fileName;
+    }
+
+    /**
+     * 通过图片资源路径，获取图片的格式
+     * @param url  图片资源路径
+     * @return  图片格式
+     */
+    private static String getFileType(String url) {
+        String[] strarr = url.split("\\.");
+        String type = strarr[strarr.length-1];
+        LogUtil.d(CLASS_NAME+"获取到的图片文件类型："+type);
+        return type;
     }
 
     /**
@@ -216,7 +253,7 @@ public class SDUtil {
      * @param height   要求高度
      * @return   缩放比例
      */
-    private static int getSampleRatio(int realWidth,int realHeight,int width,int height){
+    public static int getSampleRatio(int realWidth,int realHeight,int width,int height){
         int sampleRatio = 1;
         if(realWidth > width && realHeight > height){
             sampleRatio = Math.min(realWidth/width,realHeight/height);
